requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-appconf==1.0.4
django-extensions==2.2.9
django-extra-fields==2.0.5
django-rest-auth==0.9.5
django-sass-processor==0.8
django-sslserver==0.22
django-tagconstants==0.1
djangorestframework==3.11.0
drf-nested-routers==0.91
drf-writable-nested==0.6.0
idna==2.9
libsass==0.20.0
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wincertstore==0.2
